Move `JDACommandUser` and `ResponseEmojiMapper` to `platforms.discord.commands` package
